MAD2_Project_submission/Kanban_Project_Management/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Matplotlib==3.5.2
flask-security-too==5.0.2
flask-restful==0.3.9
redis==4.4.0
celery[redis]==5.2.7
weasyprint==57.2
Flask-Caching==2.0.1
Flask-Cors==3.0.10
